requirements.txt
python-telegram-bot
requests
chromadb
python-dotenv
beautifulsoup4